Clase Dragon actualizada, d_model

+++ src/main/java/modelo/MagoModel.java
@@ -43,7 +43,7 @@ public class MagoModel {
        }
     }
 
-    Mago leerPorId(int id){
+    Mago leerPorId(Long id){
         
         try (Session session= sessionFactory.openSession()) {
